backend/models/position.py
from extensions import db

class Position(db.Model):
    __tablename__ = 'positions'

    id = db.Column(db.Integer, primary_key=True)
    title = db.Column(db.String(100), nullable=False)

    employees = db.relationship('Employee', back_populates='position')
