backend/requirements.txt
altgraph==0.17.4
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
buildozer==1.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
crypto==1.4.1
cycler==0.12.1
distlib==0.3.9
docutils==0.21.2
et_xmlfile==2.0.0
filelock==3.17.0
filetype==1.2.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
fonttools==4.54.1
fsspec==2023.4.0
greenlet==3.2.2
httplib2==0.22.0
idna==3.10
ipaddress==1.0.23
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
keyboard==0.13.5
Kivy==2.3.1
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
Kivy-Garden==0.1.5
kivy_deps.sdl2==0.8.0
kiwisolver==1.4.7
lxml==5.3.0
MarkupSafe==2.1.3
matplotlib==3.9.2
mpmath==1.3.0
Naked==0.1.32
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pefile==2023.2.7
pexpect==4.9.0
pillow==11.0.0
ping3==4.0.8
plantuml==0.3.0
platformdirs==4.3.6
ptyprocess==0.7.0
pycryptodome==3.21.0
Pygments==2.19.1
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2025.1
pyparsing==3.2.0
PyPDF2==3.0.1
pypiwin32==223
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytube==15.0.0
pytz==2024.2
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
regex==2023.12.25
requests==2.32.3
scikit-learn==1.4.1.post1
scipy==1.12.0
setuptools==75.7.0
sh==2.2.1
shellescape==3.8.1
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.41
sympy==1.12
threadpoolctl==3.3.0
torch==2.2.0+cu118
torchaudio==2.2.0+cu118
torchvision==0.1.6
tqdm==4.66.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
virtualenv==20.29.1
Werkzeug==3.0.1
yt-dlp==2025.2.19
